// frontend/src/clients/analytics/types.gen.ts
export type ChartType = 'barchart' | 'heatmap' | 'linechart' | 'scatterplot';
// ...
export type CountType = 'document' | 'sentence' | 'mention';
// ...
export type SeriesType = 'TOPIC' | 'ANNOTATION' | 'DOCUMENT_METADATA' | 'SENTENCE_METADATA' | 'ANNOTATION_METADATA';